% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_template_methods.R
\docType{data}
\name{clires_1}
\alias{clires_1}
\alias{clires_2}
\title{CliRes dict_templates}
\format{An object of class \code{dict_template} of length 5.}
\usage{
clires_1

clires_2
}
\description{
CliRes dict_templates
}
\keyword{datasets}
