% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oucru_load_dict.R
\name{clires_auto}
\alias{clires_auto}
\title{CliRes version auto-recognition}
\usage{
clires_auto(dd_path, dd_sheets)
}
\arguments{
\item{dd_path}{path to data dictionary}

\item{dd_sheets}{sheet list of data dictionary}
}
\value{
an object of class "dict_template" or NULL
}
\description{
A function to automatically recognise which version of CliRes is the data dictionary
}
