% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fun.R
\name{reset.load.fn}
\alias{reset.load.fn}
\alias{default.load.fn}
\alias{set.default.load.fn}
\title{Backend loading functions for file extensions}
\usage{
reset.load.fn()

default.load.fn(path, ...)

set.default.load.fn(x, value)
}
\arguments{
\item{path}{path to data file, based on which will return a loading function}

\item{x}{string, file extension.}

\item{value}{a named list with two elements: fn.call for function call, fn.arg for argument passed to fn.call}
}
\value{
a function
}
\description{
set/get/reset loading functions
}
