% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{guess_datetime_order}
\alias{guess_datetime_order}
\title{Guess date-time values' order}
\usage{
guess_datetime_order(x)
}
\arguments{
\item{x}{a vector of type "character" which contains data-time values.}
}
\value{
A vector of order candidates, if no candidates, return NULL
}
\description{
Guess date-time values' order
}
