% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Smart data loading, auto assignment, and taking into account the data dictionaries}
\usage{
load_data(
  path,
  data_dict = get_dict(),
  format = TRUE,
  auto_extract = TRUE,
  log = TRUE,
  lower_case_names = FALSE,
  verbose = TRUE,
  ...,
  .loadfn = default.load.fn,
  .auto_fill_dict = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{path}{path to data file, currently support csv, xls, xlsx, and sqlite/db}

\item{data_dict}{an object containing data dict, default is call a data dictionary laid inside options, defined by load_dict()}

\item{format}{Default is TRUE: automatically format the data based on data_dict.}

\item{auto_extract}{Only respected when number of datasets within the database > 1. Default is TRUE: if the database loaded contains > 1 datasets, automatically extract it to the caller environment.}

\item{log}{Default is TRUE: generate a log containing issues regarding the data based on data dictionary.}

\item{lower_case_names}{Default is FALSE. If TRUE and auto_extract==TRUE: automatically convert dataset names to lower case}

\item{...}{additional parameters passed to respective loading function}

\item{.loadfn}{Internal, default is default.load.fn; specifying which function to use for which file extension}

\item{.auto_fill_dict}{Internal. Default is TRUE. Auto fill data dictionary based on the loaded data for variables not mentioned within the dictionaries}

\item{envir}{Internal. Default it parent.frame(): if auto_extract==TRUE, specifying which environment to extract the datasets.}

\item{vebose}{Default is TRUE: whether to print warnings, messages during loading/formatting process}
}
\value{
if auto_extract == TRUE, return TRUE, else return the loaded dataset.
}
\description{
A function to samrtly load data basing on file extensions. Data are subsequently formatted based on dictionaries.
}
\seealso{
\link{load_dict} \link{default.load.fn}
}
