% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oucru_load_dict.R, R/oucru_save_dict.R
\name{load_dict}
\alias{load_dict}
\alias{push_dict}
\alias{get_dict}
\alias{get_cat}
\alias{save_dict}
\alias{clear_dict}
\title{Load/push/remove data dictionaries}
\usage{
load_dict(path, ext, ...)

push_dict(path, ext, ...)

get_dict(dataset)

get_cat(cat)

save_dict(file)

clear_dict(dataset)
}
\arguments{
\item{path}{path to dictionary}

\item{ext}{either "rds", "xls", or "xlsx", specify format of dictionary.
If missing, will be automatically regconised using file extension.}

\item{...}{if ext == "xls" or "xlsx", addtional parameters passed to loading function.
One of them is template, default is clires_auto, which smartly selects between 2 versions of CliRes.}

\item{dataset}{which dataset'dictionary should be removed. If missing, all exisiting dictionaries are removed.}

\item{cat}{category names. If missing, print all categories.}

\item{file}{path to save dictionary}
}
\value{
TRUE
}
\description{
These functions add new dictionary, append more dictionary or remove them.
}
