% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_excel}
\alias{load_excel}
\title{Load all excel sheets}
\usage{
load_excel(path, sheets = NULL, ...)
}
\arguments{
\item{path}{path to excel file}

\item{sheets}{if NULL (default): all sheets are loaded. Otherwise, mentioned sheets are loaded}

\item{...}{additional parameters passed to readxl::read_excel}
}
\value{
a list of data.frame
}
\description{
A function to automatically load several sheets in excel files.
}
