% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_sqlite}
\alias{load_sqlite}
\title{Load all tables in sqlite files}
\usage{
load_sqlite(path, tbls = NULL, ...)
}
\arguments{
\item{path}{path to db file}

\item{tbls}{if NULL (default): all tables are loaded. Otherwise, mentioned ones are loaded}

\item{...}{addtionnal parameters passed to DBI::dbConnect}
}
\value{
a list of data frame
}
\description{
A function to automatically load several tables in sqlite database.
}
