% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oucru_load_dict.R
\name{match_val.pos}
\alias{match_val.pos}
\alias{match_pattern}
\title{Helpers for dictionary template matching functions.}
\usage{
match_val.pos(value, position = NULL, exact = TRUE)

match_pattern(pattern, perl = TRUE, ignore.case = TRUE, exact = TRUE)
}
\arguments{
\item{value}{matching pattern}

\item{position}{fall-back integer that will be return if no matched columm found.}

\item{exact}{default is TRUE, the value is matched as-is.}

\item{pattern, perl, ignore.case, exact}{parameters passed to grep}
}
\value{
for match_val.pos: a function that do the matching, for match_pattern: same as grep
}
\description{
Functions that will return a name/index of cols pointing to specific columns in data dictionaries.
}
\seealso{
\link{dict_template}
}
