% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oucru_load_dict.R, R/process_cat.R
\name{get_cat}
\alias{get_cat}
\alias{new_cat}
\alias{remove_cat}
\title{Category list manipulation}
\usage{
get_cat(cat_name)

new_cat(cat_name, cat_table)

remove_cat(cat_name)
}
\arguments{
\item{cat_name}{name of category}

\item{cat_table}{a data.frame of two columns: value<int> and label<string>, following the format of a factor}
}
\value{
TRUE
}
\description{
Add/remove/change category levels
}
