% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oucru_load_dict.R
\name{dict_template}
\alias{dict_template}
\title{Create a dictionary template}
\usage{
dict_template(...)
}
\arguments{
\item{...}{matching function for index_sheet, category_sheet, category_cols, included_cols, and data_types
See clires_1 for more information}
}
\value{
A list of class "dict_template"
}
\description{
A function to create a dict template
}
